+++ app/src/js/components/Granules/list.js
@@ -22,6 +22,7 @@ import {
   bulkActions,
   defaultWorkflowMeta,
   executeDialog,
+  groupAction,
 } from '../../utils/table-config/granules';
 import List from '../Table/Table';
 import LogViewer from '../Logs/viewer';
@@ -46,6 +47,7 @@ const AllGranules = ({
 }) => {
   const [workflow, setWorkflow] = useState(workflowOptions[0]);
   const [workflowMeta, setWorkflowMeta] = useState(defaultWorkflowMeta);
+  const [selected, setSelected] = useState([]);
   const { dropdowns } = collections;
   const { list } = granules;
   const { count, queriedAt } = list.meta;
@@ -113,7 +115,9 @@ const AllGranules = ({
         action: applyWorkflow,
       },
     };
-    return bulkActions(granules, config);
+    const selectedGranules = selected.map((id) => granules.list.data.find((g) => id === g.granuleId));
+
+    return bulkActions(granules, config, selectedGranules);
   }
 
   function selectWorkflow(selector, selectedWorkflow) {
@@ -146,6 +150,10 @@ const AllGranules = ({
     }));
   }
 
+  function updateSelection(selection) {
+    setSelected(selection);
+  }
+
   return (
     <div className="page__component">
       <section className="page__section">
@@ -169,10 +177,12 @@ const AllGranules = ({
           tableColumns={view === 'failed' ? errorTableColumns : tableColumns}
           query={query}
           bulkActions={generateBulkActions()}
+          groupAction={groupAction}
           rowId="granuleId"
           initialSortId={tablesortId}
           filterAction={filterGranules}
           filterClear={clearGranulesFilter}
+          onSelect={updateSelection}
         >
           <ListFilters>
             <Dropdown