+++ app/src/js/utils/table-config/granules.js
@@ -22,6 +22,7 @@ import {
   reingestGranuleClearError,
   removeGranule,
   removeGranuleClearError,
+  removeAndDeleteGranule
 } from '../../actions';
 import ErrorReport from '../../components/Errors/report';
 import { strings } from '../../components/locale';
@@ -31,6 +32,12 @@ import BatchReingestConfirmContent from '../../components/ReingestGranules/Batch
 import BatchReingestCompleteContent from '../../components/ReingestGranules/BatchReingestCompleteContent';
 import TextArea from '../../components/TextAreaForm/text-area';
 import { getPersistentQueryParams, historyPushWithQueryParams } from '../url-helper';
+import GranuleInventory from '../../components/Granules/granule-inventory';
+
+export const groupAction = {
+  title: 'Granule Actions',
+  description: 'Select the action you would like to perform on the selected granules from the table below',
+};
 
 export const tableColumns = [
   {
@@ -170,6 +177,7 @@ const confirmReingest = (d) => `Reingest ${d} Granule${d > 1 ? 's' : ''}?`;
 const confirmApply = (d) => `Run workflow on ${d} granules?`;
 const confirmRemove = (d) => `Remove ${d} granule(s) from ${strings.cmr}?`;
 const confirmDelete = (d) => `Delete ${d} granule(s)?`;
+const confirmRemoveFromCMR = (d) => 'Selection contains granules that are published to CMR which must be removed before deleting. Remove published granules from CMR and delete?';
 
 /**
  * Determine the base context of a collection view
@@ -242,6 +250,20 @@ const granuleModalJourney = ({
   return modalOptions;
 };
 
+const containsPublishedGranules = (selectedGranules) => {
+  let publishedGranules = [];
+
+  if (Array.isArray(selectedGranules) && selectedGranules.length > 0) {
+    publishedGranules = selectedGranules.filter((g) => g.published === true);
+  }
+
+  if (publishedGranules.length < 1) {
+    return false;
+  }
+
+  return true;
+};
+
 export const reingestAction = (granules) => ({
   text: 'Reingest',
   action: reingestGranule,
@@ -252,7 +274,7 @@ export const reingestAction = (granules) => ({
   getModalOptions: granuleModalJourney
 });
 
-export const bulkActions = (granules, config) => [
+export const bulkActions = (granules, config, selectedGranules) => [
   reingestAction(granules),
   {
     text: 'Execute',
@@ -262,6 +284,14 @@ export const bulkActions = (granules, config) => [
     confirm: confirmApply,
     confirmOptions: config.execute.options,
     className: 'button--execute'
+  },
+  {
+    Component:
+      <GranuleInventory
+        element='button'
+        className='csv__download button button--small button--file button--green form-group__element'
+        confirmAction={true}
+      />
   },
   {
     text: strings.remove_from_cmr,
@@ -281,9 +311,9 @@ export const bulkActions = (granules, config) => [
   },
   {
     text: 'Delete',
-    action: deleteGranule,
+    action: containsPublishedGranules(selectedGranules) ? removeAndDeleteGranule : deleteGranule,
     state: granules.deleted,
     clearError: deleteGranuleClearError,
-    confirm: confirmDelete,
+    confirm: containsPublishedGranules(selectedGranules) ? confirmRemoveFromCMR : confirmDelete,
     className: 'button--delete'
   }];