+++ cypress/integration/granules_spec.js
@@ -145,14 +145,6 @@ describe('Dashboard Granules Page', () => {
       );
     });
 
-    it('should display a link to download the granule list', () => {
-      cy.visit('/granules');
-
-      cy.contains('.heading--xlarge', 'Granules');
-
-      cy.contains('a', 'Create Granule Inventory List');
-    });
-
     it('Should update dropdown with label when visiting bookmarkable URL', () => {
       cy.visit('/granules?status=running');
       cy.get('.filter-status .rbt-input-main').as('status-input');
@@ -310,7 +302,8 @@ describe('Dashboard Granules Page', () => {
       cy.route('GET', `/granules/${granuleId}`).as('getGranule');
       cy.visit('/granules');
       cy.get(`[data-value="${granuleId}"] > .td >input[type="checkbox"]`).click();
-      cy.get('.list-actions').contains('Reingest').click();
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Reingest').click();
       cy.get('.button--submit').click();
       cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Success');
       cy.get('.button__goto').click();
@@ -334,7 +327,8 @@ describe('Dashboard Granules Page', () => {
       cy.visit('/granules');
       cy.get(`[data-value="${granuleIds[0]}"] > .td >input[type="checkbox"]`).click();
       cy.get(`[data-value="${granuleIds[1]}"] > .td >input[type="checkbox"]`).click();
-      cy.get('.list-actions').contains('Reingest').click();
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Reingest').click();
       cy.get('.button--submit').click();
       cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Success');
       cy.get('.button__goto').click();
@@ -357,7 +351,8 @@ describe('Dashboard Granules Page', () => {
       cy.visit('/granules');
       cy.get(`[data-value="${granuleIds[0]}"] > .td >input[type="checkbox"]`).click();
       cy.get(`[data-value="${granuleIds[1]}"] > .td >input[type="checkbox"]`).click();
-      cy.get('.list-actions').contains('Reingest').click();
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Reingest').click();
       cy.get('.button--submit').click();
       cy.get('.modal-content .modal-body .alert', { timeout: 10000 }).should('contain.text', 'Error');
       cy.get('.Collapsible__contentInner').should('contain.text', 'Oopsie');
@@ -420,7 +415,8 @@ describe('Dashboard Granules Page', () => {
       cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
       cy.get('@granule-checkbox').click({ multiple: true });
 
-      cy.get('.csv__download').click();
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Create Granule Inventory List').click();
       cy.get('.default-modal.granule-inventory ').as('modal');
 
       cy.get('@modal').contains('div', 'You have generated a selection to process for the following list:');
@@ -541,6 +537,146 @@ describe('Dashboard Granules Page', () => {
           .should('contain', 0)
           .and('contain', 'Running');
       });
+    });
+
+    it('Should show the correct DELETE modal when published granules ARE selected', () => {
+      // this granules query will have a combination of published and unpublished granules
+      cy.visit('/granules?limit=50&page=1&status=completed');
+
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Delete').click();
+      cy.get('.default-modal.batch-async-modal ').as('modal');
+
+      cy.get('@modal').contains('div', 'Selection contains granules that are published to CMR which must be removed before deleting. Remove published granules from CMR and delete?');
+    });
+
+    it('Should show the correct DELETE modal when published granules are selected on the "processing granules" page', () => {
+      // All granules in the "processing" state are published
+      cy.visit('/granules/processing?limit=50');
+
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Delete').click();
+      cy.get('.default-modal.batch-async-modal ').as('modal');
+
+      cy.get('@modal').contains('div', 'Selection contains granules that are published to CMR which must be removed before deleting. Remove published granules from CMR and delete?');
+    });
+
+    it('Should show the correct DELETE modal when published granules are NOT selected', () => {
+      // published = false will return a table with only unpublished granules
+      cy.visit('/granules?limit=50&page=1&published=false');
+
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Delete').click();
+      cy.get('.default-modal.batch-async-modal ').as('modal');
+
+      cy.get('@modal').contains('div', 'Delete 5 granule(s)?');
+    });
+
+    it('Should handle a successful API response from the Remove and Delete granule requests', () => {
+      cy.server();
+      cy.route({
+        method: 'PUT',
+        url: '/granules/*',
+        status: 200,
+        response: { message: 'success' }
+      });
+
+      cy.route({
+        method: 'DELETE',
+        url: '/granules/*',
+        status: 200,
+        response: { message: 'success' }
+      });
+
+      // this granules query will have a combination of published and unpublished granules
+      cy.visit('/granules?limit=50&page=1&status=completed');
+
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Delete').click();
+      cy.contains('.button--submit', 'Confirm').click();
+
+      cy.route2({
+        url: '/granules',
+        method: 'PUT'
+      }, (req) => {
+        const requestBody = JSON.parse(req.body);
+        expect(requestBody).to.have.property('action', 'removeFromCmr');
+      });
+
+      cy.get('.default-modal.batch-async-modal ').as('modal');
+      cy.get('@modal').contains('div', 'Success!');
+    });
+
+    it('Should handle a failed API response from the Remove granule requests', () => {
+      cy.server();
+      cy.route({
+        method: 'PUT',
+        url: '/granules/*',
+        status: 400,
+        response: { message: 'error' }
+      });
+
+      cy.route({
+        method: 'DELETE',
+        url: '/granules/*',
+        status: 200,
+        response: { message: 'success' }
+      });
+
+      // this granules query will have a combination of published and unpublished granules
+      cy.visit('/granules?limit=50&page=1&status=completed');
+
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Delete').click();
+      cy.contains('.button--submit', 'Confirm').click();
+
+      cy.get('.default-modal.batch-async-modal ').as('modal');
+      cy.get('@modal').contains('div', 'Error');
+    });
+
+    it('Should handle a failed API response from the Delete granule requests', () => {
+      cy.server();
+      cy.route({
+        method: 'PUT',
+        url: '/granules/*',
+        status: 200,
+        response: { message: 'success' }
+      });
+
+      cy.route({
+        method: 'DELETE',
+        url: '/granules/*',
+        status: 400,
+        response: { message: 'error' }
+      });
+
+      // this granules query will have a combination of published and unpublished granules
+      cy.visit('/granules?limit=50&page=1&status=completed');
+
+      cy.get('.table .tbody .tr .td input[type=checkbox]').as('granule-checkbox');
+      cy.get('@granule-checkbox').click({ multiple: true });
+
+      cy.contains('button', 'Granule Actions').click();
+      cy.contains('button', 'Delete').click();
+      cy.contains('.button--submit', 'Confirm').click();
+
+      cy.get('.default-modal.batch-async-modal ').as('modal');
+      cy.get('@modal').contains('div', 'Error');
     });
   });
 });